+++ smsplus/shared.h
@@ -37,6 +37,7 @@ extern "C" {
 #include "sms.h"
 #include "vdp.h"
 #include "render.h"
+#include "tms.h"
 #include "sn76496.h"
 #include "system.h"
 extern int soundEnabled;